debug cpu rate both process and processor

+++ include/processor.h
@@ -10,8 +10,8 @@ class Processor {
 
  // TODO: (Done) Declare any necessary private members
  private:
- long PrevTotal{0};
- long PrevIdle{0};
+ float PrevTotal{0.0f};
+ float PrevIdle{0.0f};
 };
 
 #endif
